third_party/WebKit/LayoutTests/external/wpt/xhr/headers-normalize-response-expected.txt

This is a testharness.js-based test.
FAIL Header value: hello_world\0 assert_equals: expected "hello world\0" but got "hello world"
FAIL Header value: \0hello_world assert_equals: expected "\0hello world" but got "hello world"
FAIL Header value: hello\0world assert_equals: expected "hello\0world" but got "helloworld"
PASS Header value: __hello_world
PASS Header value: hello_world__
PASS Header value: __hello_world__
PASS Header value: [tab]hello_world
PASS Header value: hello_world[tab]
PASS Header value: [tab]hello_world[tab]
PASS Header value: hello______world
PASS Header value: hello[tab]world
FAIL Header value: \0 assert_equals: expected "\0" but got ""
PASS Header value: ___
PASS Header value: [tab]
PASS Header value: 
Harness: the test ran to completion.
